+++ test/runtests.jl
@@ -40,7 +40,7 @@ include("common.jl")
 
     include("wf.jl")
 
-    #include("owf.jl")
+    include("owf.jl")
 
     #include("des.jl")