+++ CHANGELOG.md
@@ -5,6 +5,7 @@ WaterModels.jl Change Log
 - Decoupled `check_valve` and `shutoff_valve` objects from `pipe` objects.
 - Introduced a new `valve` component.
 - Renamed `pressure_reducing_valve` to `regulator`.
+- Decomposed component constraints a bit more.
 
 ### v0.5.0
 - Rename formulation types.